[v] purchase order > delivered should able create GRN and vendor bill after GRN. [v] vendor bill post should create bill number [v] vendor bill should be able to create payment [v] posting vendor bill should create entry [v] cancelling vendor bill should reverse entry. [v] post customer invoice should create entry [v] cancelling customer invoice should reverse entry. [v] sales receipt / CI > post without cost basis should result in error [v] purchase receipt > post should create entries. [v] purchase receipt > cancel should revert entries. [v] to apply rollback transaction functionality for nextval sequence function [v] Product Moving Average doesn't change after adding Purchase Receipt

+++ Nuggets.Application/Services/CustomerInvoiceService.cs
@@ -125,7 +125,7 @@ public sealed class CustomerInvoiceService(
                 (string.IsNullOrEmpty(existing.InvoiceNumber) || existing.InvoiceNumber.StartsWith("Draft CI")))
             {
                 // Generate auto number
-                var nextNumber = await repo.GetNextSequenceValueAsync("customer_invoice_number_seq");
+                var nextNumber = await repo.GetNextSequenceValueAsync("customer_invoice_number_seq", tx);
                 existing.InvoiceNumber = $"CI/{dto.InvoiceDate.Year}/{nextNumber:000000}";
             }
             
@@ -207,6 +207,12 @@ public sealed class CustomerInvoiceService(
         {
             var existing = await repo.GetByIdAsync(id);
             if (existing is null) return Result<bool>.Err("Invoice not found", "NOT_FOUND");
+
+            if (!(string.IsNullOrEmpty(existing.InvoiceNumber) || existing.InvoiceNumber.StartsWith("Draft CI")))
+            {
+                return Result<bool>.Err("You are not allowed to delete a transaction with existing number.", "VALIDATION_ERROR");
+            }
+            
             await repo.DeleteAsync(existing);
             await tx.CommitAsync();
             return Result<bool>.Ok(true);
